Filter button base + css + refactor all js
