+++ FrontEnd/index.html
@@ -9,7 +9,7 @@
 	<link href="https://fonts.googleapis.com/css2?family=Syne:wght@700;800&family=Work+Sans&display=swap" rel="stylesheet">
 	<meta name="description" content="">
 	<link rel="stylesheet" href="./assets/style.css">
-	<script src="JavaScript/main.js" type="module"></script>
+	<script src="JavaScript/main.js" type="module" defer></script>
 </head>
 <body>
 <header>
